Programatically build the color on the timeline
Using the color library we loop through the various steps in the
timeline and add color with a programmatic gradient. This means that we
won't need to hand code these elements in the css and gives us greater
flexibility changing out the data and time period intervals in the
future.