New method for getting the current era interval
We now use the current era in the collection based on the xPos of the
mouse to select which map to show. The interval is set off of the i
value in the model.

The hard part of this code was figuring out how to find a model from a
collection based on the attribute in the model. The key was the filter
method and looking for a xPos that was smaller than the era's right
boundary.